Changed Package.py to Installer.py
